Allow rzls to launch using standard dotnet (Redo)
- Updates where to pick up dotnet.exe.
- Contributes to making rzls move away from being self-contained.
- Unifies how roslyn and razor servers acquire dotnet path/env variables.
- Picks up leftover changes from https://github.com/dotnet/vscode-csharp/pull/5855 into main
