+++ src/lsptoolshost/roslynLanguageServer.ts
@@ -491,29 +491,9 @@ export class RoslynLanguageServer {
         const serverPath = getServerPath(platformInfo);
 
         const dotnetInfo = await hostExecutableResolver.getHostExecutableInfo();
-        const dotnetRuntimePath = path.dirname(dotnetInfo.path);
         const dotnetExecutablePath = dotnetInfo.path;
 
         _channel.appendLine('Dotnet path: ' + dotnetExecutablePath);
-
-        // Take care to always run .NET processes on the runtime that we intend.
-        // The dotnet.exe we point to should not go looking for other runtimes.
-        const env: NodeJS.ProcessEnv = { ...process.env };
-        env.DOTNET_ROOT = dotnetRuntimePath;
-        env.DOTNET_MULTILEVEL_LOOKUP = '0';
-        // Save user's DOTNET_ROOT env-var value so server can recover the user setting when needed
-        env.DOTNET_ROOT_USER = process.env.DOTNET_ROOT ?? 'EMPTY';
-
-        if (languageServerOptions.crashDumpPath) {
-            // Enable dump collection
-            env.DOTNET_DbgEnableMiniDump = '1';
-            // Collect heap dump
-            env.DOTNET_DbgMiniDumpType = '2';
-            // Collect crashreport.json with additional thread and stack frame information.
-            env.DOTNET_EnableCrashReport = '1';
-            // The dump file name format is <executable>.<pid>.dmp
-            env.DOTNET_DbgMiniDumpName = path.join(languageServerOptions.crashDumpPath, '%e.%p.dmp');
-        }
 
         let args: string[] = [];
 
@@ -557,7 +537,7 @@ export class RoslynLanguageServer {
             const csharpDevKitArgs = this.getCSharpDevKitExportArgs();
             args = args.concat(csharpDevKitArgs);
 
-            await this.setupDevKitEnvironment(env, csharpDevkitExtension);
+            await this.setupDevKitEnvironment(dotnetInfo.env, csharpDevkitExtension);
         } else {
             // C# Dev Kit is not installed - continue C#-only activation.
             _channel.appendLine('Activating C# standalone...');
@@ -580,7 +560,7 @@ export class RoslynLanguageServer {
         const cpOptions: cp.SpawnOptionsWithoutStdio = {
             detached: true,
             windowsHide: true,
-            env: env,
+            env: dotnetInfo.env,
         };
 
         if (serverPath.endsWith('.dll')) {
